refactor: separate pipeline builder tests
